pyproject.toml
[project]
name = "subtitle-converter"
version = "0.1.0"
description = "Web app to search, download and translate subtitles from OpenSubtitles.org"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "pysrt>=1.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]
